+++ src/api/domain/decks/deck.controller.ts
@@ -2,9 +2,11 @@ import { Controller } from '@api/controller';
 import { Router, RequestHandler } from 'express';
 import { createDeckActionValidation } from './actions/create-deck.action';
 import { authorizationMiddleware } from '@api/middlewares/auth/auth.middleware';
+import { multerUpload } from './actions/upload-deck-image.action';
 
 interface Dependencies {
   createDeckAction: RequestHandler;
+  uploadDeckImageAction: RequestHandler;
 }
 
 export class DeckController extends Controller {
@@ -19,6 +21,12 @@ export class DeckController extends Controller {
       '/',
       [authorizationMiddleware, createDeckActionValidation],
       this.dependencies.createDeckAction,
+    );
+
+    router.post(
+      '/upload-image',
+      [multerUpload.single('file')],
+      this.dependencies.uploadDeckImageAction,
     );
 
     return router;